Estimation and Copy report set in the same script (create-html-report.ts).
